* Move metadata API spi file out of dubbo-config-api
