+++ dubbo-registry/dubbo-registry-api/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.common.deploy.ApplicationDeployListener
@@ -1,1 +1,1 @@
-exporter=org.apache.dubbo.registry.deploy.ExporterDeployListener
+exporter=org.apache.dubbo.config.metadata.ExporterDeployListener
