+++ dubbo-registry/dubbo-registry-api/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.registry.client.ServiceInstanceCustomizer
@@ -1,4 +1,4 @@
 protocol-ports=org.apache.dubbo.registry.client.metadata.ProtocolPortsMetadataCustomizer
 instance-metadata=org.apache.dubbo.registry.client.metadata.ServiceInstanceMetadataCustomizer
 port=org.apache.dubbo.registry.client.metadata.ServiceInstanceHostPortCustomizer
-metadata-url=org.apache.dubbo.registry.deploy.MetadataServiceURLParamsMetadataCustomizer
+metadata-url=org.apache.dubbo.config.metadata.MetadataServiceURLParamsMetadataCustomizer
